PTE_ASR_GRAMMAR_DOCKER/requirements.txt

language-tool-python
fastapi
uvicorn
python-multipart
requests
openai-whisper
torch
